requirements.txt
streamlit
openai
python-dotenv
PyMuPDF
pydantic
SQLAlchemy
pika